requirements.txt
instabot==0.117.0
python-dotenv==0.9.1
requests==2.23.0
